+++ sbi/utils/metrics.py
@@ -522,6 +522,44 @@ def _test():
     # unbiased_mmd_squared_hypothesis_test(x, y)
 
 
+def l2(x: Tensor, y: Tensor, axis=-1) -> Tensor:
+    """
+    Calculates the L2 distance between two tensors. Note, we cannot use the
+    torch.nn.MSELoss function as this sums across the batch dimension AND the
+    dimension given by <axis>. For tarp, we only require to sum across
+    the <axis> dimension.
+
+    Args:
+        x (Tensor): The first tensor.
+        y (Tensor): The second tensor.
+        axis (int, optional): The axis along which to calculate the L2 distance.
+                Defaults to -1.
+    Returns:
+        Tensor: A tensor containing the L2 distance between x and y along the
+                specified axis.
+    """
+    return torch.sqrt(torch.sum((x - y) ** 2, axis=axis))
+
+
+def l1(x: Tensor, y: Tensor, axis=-1) -> Tensor:
+    """
+    Calculates the L1 distance between two tensors. Note, we cannot use the
+    torch.nn.L1Loss function as this sums across the batch dimension AND the
+    dimension given by <axis>. For tarp, we only require to sum across
+    the <axis> dimension.
+
+    Args:
+        x (Tensor): The first tensor.
+        y (Tensor): The second tensor.
+        axis (int, optional): The axis along which to calculate the L1 distance.
+                Defaults to -1.
+    Returns:
+        Tensor: A tensor containing the L1 distance between x and y along the
+                specified axis.
+    """
+    return torch.sum(torch.abs(x - y), axis=axis)
+
+
 def main():
     _test()
